split test to separate files for compilation speed. add some missing includes.
