+++ c++-src/inc/cgui/warnings.hpp
@@ -1,6 +1,3 @@
-//
-// Created by rvons on 2024-07-09.
-//
 
 #ifndef CGUI_CGUI_WARNINGS_HPP
 #define CGUI_CGUI_WARNINGS_HPP
